Refactor user auth pages
